Chpter 5: Fantasy Game Inventory
I am creating a fantasy game inventory. Data structure to model the inventory will be a dictionary where the keys are string values describing the item in the inventory and the value is an integer value detailing how many of that item the player has. I am to write a function named "displayInventory()" that would take any possible "inventory" and display according to the parameters defined in the automatetheboringstuff.com/chapter5/ Practice Projects.